requirements.txt
Flask==3.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
python-decouple==3.8
openai>=1.40.0
